Wire Batch.plot through cellpy.plotting and delete batch_plotters.
Relocate the cycle-life summary path into plotting, triage backends
(plotly/matplotlib; seaborn alias; bokeh removed), and keep the
b.plotter.figure facade so batch notebooks stay usable.
