+++ docs/api/plotting.md
@@ -2,9 +2,10 @@
 
 The shared plotting machinery. Figure loading and saving, legend and marker
 post-processing, and the plotly templates all live here in **one** copy —
-`cellpy.utils.plotutils`, `cellpy.utils.collectors` and
-`cellpy.utils.batch_tools.batch_plotters` re-export from it, so existing
-imports keep working.
+`cellpy.utils.plotutils` and `cellpy.utils.collectors` re-export from it, so
+existing imports keep working. `Batch.plot` delegates to
+`cellpy.plotting.batch_summary_plot` (#658); the old
+`cellpy.utils.batch_tools.batch_plotters` module is gone.
 
 The drawing functions themselves (`summary_plot`, `raw_plot`,
 `cycle_info_plot`, `cycles_plot`) still live in
@@ -15,3 +16,5 @@ The drawing functions themselves (`summary_plot`, `raw_plot`,
 ::: cellpy.plotting.labels
 
 ::: cellpy.plotting.theme
+
+::: cellpy.plotting.batch_summary
